feat: add view as boilerplate code
