+++ src/app/Scenes/MyCollection/Components/MyCollectionCollectedArtists.tsx
@@ -1,17 +1,15 @@
 import { Text, useSpace } from "@artsy/palette-mobile"
-import { MyCollectionCollectedArtistsRail_me$key } from "__generated__/MyCollectionCollectedArtistsRail_me.graphql"
-import { MyCollectionCollectedArtistsOnlyView } from "app/Scenes/MyCollection/Components/MyCollectionCollectedArtistsOnlyView"
+import { MyCollectionCollectedArtists_me$key } from "__generated__/MyCollectionCollectedArtists_me.graphql"
+import { MyCollectionCollectedArtistsView } from "app/Scenes/MyCollection/Components/MyCollectionCollectedArtistsView"
 import { MyCollectionTabsStore } from "app/Scenes/MyCollection/State/MyCollectionTabsStore"
 import { ScrollView } from "react-native"
 import { graphql, useFragment } from "react-relay"
 
-interface MyCollectionCollectedArtistsRail {
-  me: MyCollectionCollectedArtistsRail_me$key
+interface MyCollectionCollectedArtists {
+  me: MyCollectionCollectedArtists_me$key
 }
 
-export const MyCollectionCollectedArtistsRail: React.FC<MyCollectionCollectedArtistsRail> = ({
-  me,
-}) => {
+export const MyCollectionCollectedArtists: React.FC<MyCollectionCollectedArtists> = ({ me }) => {
   const space = useSpace()
   const selectedTab = MyCollectionTabsStore.useStoreState((state) => state.selectedTab)
   const data = useFragment(collectedArtistsFragment, me)
@@ -21,7 +19,7 @@ export const MyCollectionCollectedArtistsRail: React.FC<MyCollectionCollectedArt
   }
 
   if (selectedTab === "Artists") {
-    return <MyCollectionCollectedArtistsOnlyView me={data} />
+    return <MyCollectionCollectedArtistsView me={data} />
   }
 
   return (
@@ -32,7 +30,7 @@ export const MyCollectionCollectedArtistsRail: React.FC<MyCollectionCollectedArt
 }
 
 const collectedArtistsFragment = graphql`
-  fragment MyCollectionCollectedArtistsRail_me on Me {
-    ...MyCollectionCollectedArtistsOnlyView_me
+  fragment MyCollectionCollectedArtists_me on Me {
+    ...MyCollectionCollectedArtistsView_me
   }
 `
